requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-social-oauth2==1.2.1
drf-yasg==1.20.0
ecdsa==0.17.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
oauthlib==3.2.0
packaging==21.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
